refactor(backend): move `TaskController` to `libs/backend`
We're going to use it in two apps.
