+++ apps/scan/backend/src/electrical_testing/context.ts
@@ -1,11 +1,10 @@
 import { InsertedSmartCardAuthApi } from '@votingworks/auth';
+import { TaskController } from '@votingworks/backend';
 import { Logger } from '@votingworks/logging';
 import { UsbDrive } from '@votingworks/usb-drive';
-
 import { Printer } from '../printing/printer';
 import { Workspace } from '../util/workspace';
 import { SimpleScannerClient } from './simple_scanner_client';
-import { TaskController } from './task_controller';
 
 export interface ServerContext {
   auth: InsertedSmartCardAuthApi;
