+++ libs/backend/src/index.ts
@@ -15,6 +15,7 @@ export * from './pdf_to_text';
 export * from './system_call';
 export * from './scan_globals';
 export * from './split';
+export * from './task_controller';
 export * from './ui_strings';
 export * from './language_and_audio';
